divmod

+++ divmod.go
@@ -2,7 +2,7 @@ package piscine
 
 func DivMod(a, b int, div, mod *int) {
 
-	с := int(a / b) // можно просто a/b
+	с := (a / b)
 	d := a % b
 	*div = c
 	*mod = d
